Feat: Add error handling logic and some Snack(bar)s

+++ lib/service/household_service.dart
@@ -58,20 +58,24 @@ class HouseholdService {
     required String userId,
     required String name,
   }) async {
-    DocumentReference ref = FirebaseFirestore.instance.doc('User/$userId');
+    try {
+      DocumentReference ref = FirebaseFirestore.instance.doc('User/$userId');
 
-    String uuid = _uuid.v4();
-    String groupId =
-        sha256.convert(utf8.encode(uuid)).toString().substring(0, 8);
+      String uuid = _uuid.v4();
+      String groupId =
+          sha256.convert(utf8.encode(uuid)).toString().substring(0, 8);
 
-    DocumentReference newId = await _householdRepository.add(Household(
-      id: 'placeholder',
-      admin: ref,
-      name: name,
-      inhabitants: [ref],
-      groupId: groupId,
-    ));
-    return newId;
+      DocumentReference newId = await _householdRepository.add(Household(
+        id: 'placeholder',
+        admin: ref,
+        name: name,
+        inhabitants: [ref],
+        groupId: groupId,
+      ));
+      return newId;
+    } catch (e) {
+      throw Exception('Failed to create household: $e');
+    }
   }
 
   Future<void> addTaskToHousehold(
@@ -126,18 +130,30 @@ class HouseholdService {
       throw Exception('No household found.');
     }
 
-    // Remove inhabitant from inhabitants list of a household
     final DocumentReference userRef =
         FirebaseFirestore.instance.doc('User/$userId');
+    if (!targetHousehold.inhabitants.contains(userRef)) {
+      throw Exception('User not found in household.');
+    }
+
+    // Remove inhabitant from inhabitants list of a household
     targetHousehold.inhabitants.remove(userRef);
-    _householdRepository.updateEntity(targetHousehold.id, targetHousehold);
+    await _householdRepository.updateEntity(
+        targetHousehold.id, targetHousehold);
 
     // Remove household from list of households of an inhabitant
     final InhabitantService inhabitantService =
         GetIt.instance<InhabitantService>();
     final DocumentReference householdRef =
         FirebaseFirestore.instance.doc('Household/${targetHousehold.id}');
-    inhabitantService.removeHouseholdFromInhabitant(
+    Inhabitant? user = await inhabitantService.getInhabitant(userId);
+    if (user == null) {
+      throw Exception('User not found.');
+    }
+    if (!user.households.contains(householdRef)) {
+      throw Exception('Household not found in user\'s list.');
+    }
+    await inhabitantService.removeHouseholdFromInhabitant(
       uid: userId,
       newRef: householdRef,
     );
